TransDualNet/requirements.txt
Pillow~=9.3.0
torch~=2.1.0+cu118
numpy~=1.23.5
scipy~=1.10.1
matplotlib~=3.6.2
tqdm~=4.64.1
torchvision~=0.16.0+cu118
scikit-learn~=1.3.2